Add LICENCE.md to the template. (#26)
Solves #19.

We had two options.

1. As done in
[audreyfeldroy](https://github.com/audreyfeldroy)/[cookiecutter-pypackage](https://github.com/audreyfeldroy/cookiecutter-pypackage):
a big if-elseif-elseif block containing the whole licence text.
2. As done in
[neuroinformatics-unit](https://github.com/neuroinformatics-unit)/[python-cookiecutter](https://github.com/neuroinformatics-unit/python-cookiecutter):
a small if-elseif-elseif block that copies over licences from a
directory (a selection are committed to the repo). But then one has to
[delete the licences directory in a post-gen
hook](https://github.com/neuroinformatics-unit/python-cookiecutter/blob/main/hooks/post_gen_project.py).

I opted for option 1. And went with a nicely markdown-ed version of
GPL-3.0 courtesy of
[IQAndreas](https://github.com/IQAndreas)/[markdown-licenses](https://github.com/IQAndreas/markdown-licenses).

Linter is expected to fail because the internal template's
pyproject.toml has some illegal toml.