Always ignore template at top level pre-commit (#320)
We lint at the top level, and at the template level. It doesn't really
make sense to template both, and some of the hooks get confused with the
`cookiecutter` templating. Previously, we were ignoring some, but
actually makes sense to ignore it all.

The downside is that locally it will be annoying to lint at the package
level. But that is fine because CI with lint at the template level.
Sometimes you may need to create a dummy project to fix the issue.