Start a utility libraries page. (#328)
Start a utilities page rather than solving #304 as-requested.

As I said in the ticket: I'm 👎 about adding these kinds of dependencies
to our template. Not least because not everyone _needs_ a logger, but
also because these choices feel out-of-scope. We probably _should_
highlight libraries that we like and prefer to the builtins.

I've added my opinions, such as they are. Challenges are welcome.

## Closes
- #304.

---------

Co-authored-by: Patrick J. Roddy <patrickjamesroddy@gmail.com>