Add CITATION.cff file (#340)
Part of #330 

Adds a CITATION.cff file for this repository with all current
contributors included as authors.
The ordering of authors should match the non-bot accounts on
https://github.com/UCL-ARC/python-tooling/graphs/contributors

I have added ORCIDs for everyone I could find them for either from
GitHub profile or ARC people directory. Similarly I have tried to enter
given and family names as listed on GitHub Profile or ARC people
directory but add a review comment if I've made any mistakes or you
would prefer to be listed differently.

I am requesting a review from everyone who is listed as an author. If
you are happy to be included as an author and don't want any changes
made to your information please approve. If you need any changes made to
your information or want to be removed as an author please add a review
and request changes accordingly.

As I ran out of reviewer slots (maximum of 15 😅) I'm manually tagging
@paddyroddy and @samcunliffe

---------

Co-authored-by: Patrick J. Roddy <patrickjamesroddy@gmail.com>