Drop `3.9` support (#360)
Following the recent change on GitHub Actions
https://github.com/actions/setup-python/issues/850 our tests are
currently failing https://github.com/UCL-ARC/python-tooling/issues/359.

Previously we were following
[NEP29](https://numpy.org/neps/nep-0029-deprecation_policy.html) which
is why we never supported `3.8`. My suggestion is to drop `3.9` which
means nothing else has to change.

The alternative is to do @dstansby's
[suggestion](https://github.com/UCL-ARC/python-tooling/issues/359#issuecomment-2077127044),
but I think this is low-key complicated when involving our `tox` setup
too. Personally, I don't think it's worth it, and we should fix the
tests ASAP.