Create "Recommended libraries" category page with subpages. (#428)
As I'm one the who was blocking #373 and thumbs down-ing all over the
place, here's a proposal to resolve #427.

## Resolves
- #427 

## Follows on from
- #328 
- #373 

## tl;dr of changes
This PR:
* Adds a parent "Recommended libraries" page with children: guis,
parallel and async.
* Split's the non-obvious "Utility libraries" into clis and logging and
moves them also under the parent.
* Adds some blurb to the parent page itself, which should be scrutinised
for clarity and readability please!

## Choices:
I decide "Recommended libraries" rather than "Recommended modules" or
"Utlility <anything>". Because a) some things in my brain's definition
of the tools/not tools split are Python modules but still tool-ish and
b) modules is a Python name but libraries feels _relatively_ language
agnostic (to me, at least). I'm thinking of newcomers to Python.