Add release workflow
This will create a `latest` tag that will track the `release` tag,
so that we can encourage users to use the `latest` tag, but when
submitting a bug report we could use the actual version.

Note if you don't follow SemVer then this wouldn't work. But that is
kind of unavoidable.