Rework README and move dev instructions to contributing (#481)
I've attempted to simplify the user-facing quick-start instructions. Not
a huge fan of enumerated instructions and the optional (but cool)
one-liner it doesn't really fit in 1. 2. 3....

Also it's a bit overkill to have a venv/conda environment just for
cookiecutter itself. The cookiecutter docs recommend `pipx install`. I
claim that we want to encourage `uvx` because it's fast and cool 😎 .

I've spent a bit of effort trying to be clear but comments on clarity
etc are very welcome.

I also moved and merged the dev instructions into CONTRIBUTING so README
is _**just**_ user-facing. And CONTRIBUTING is just dev.

This addresses
- #466 

And two comments
-
https://github.com/UCL-ARC/python-tooling/pull/464#discussion_r1810505958
-
https://github.com/UCL-ARC/python-tooling/issues/419#issuecomment-2434644366

---------

Co-authored-by: David Stansby <dstansby@gmail.com>
Co-authored-by: Patrick J. Roddy <patrickjamesroddy@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>