Improve accessibility of traffic lights: replacing 🟢🟠🔴 with JustTheDoc's labels. (#531)
An attempt to solve

- #526 

by replacing the emoji (same shape different colours) with a
[JustTheDocs
label](https://just-the-docs.com/docs/ui-components/labels/) UI
component.
Unfortunately, the JTD-flavoured markdown syntax for this is 

```md
Good
{: .label .label-green}
```

... that is to say: it needs two lines. And obviously, this doesn't work
inside markdown tables.

But once upon a time, I learned some HTML 🙃 so we can also access them
as
```html
<span class="foo"></span>
```
 tags.

----

Now! This makes the markdown not look as beautiful/colourful/emojiful
when viewed in GitHub...

![Screenshot 2025-04-11 at 16 07
24](https://github.com/user-attachments/assets/ba17275a-6169-40ca-b481-009adc5dffce)

(though arguably more readable... which is perhaps the accessibility
point.)

----

Here's what it looks like on the built pages:

![Screenshot 2025-04-11 at 16 04
10](https://github.com/user-attachments/assets/227dcafa-2e8f-4385-9595-f10807796ebb)

![Screenshot 2025-04-11 at 16 04
17](https://github.com/user-attachments/assets/893e1d5e-0445-4838-946a-8a6ab9ec85a6)

---------

Co-authored-by: Patrick J. Roddy <patrickjamesroddy@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>