Stop `vale` failing for forked contributions that don't touch prose. (#547)
Relates to 
- #549

(i.e. the prose lint failure for pull requests from forked
repositories).
Should improve the quality of life of external contributing devs.

This change _only_ annotates the diff introduced by a PR. So #549 will
still be a bug for external contributors who write a lot of prose. But
for technical contributions, at least, we avoid the 10 annotations
limit.