Simplify install instructions (#555)
My 'hot' take is that the goal of the install instructions should be a
single direct recommended way to use the template, with minimal
diversions. Here I've changed the alternative install instructions to
still link to the `cookiecutter` install instructions, but to shorten it
to reduce potential diversion from the recommended way.